% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_hnsw}
\alias{add_hnsw}
\title{Add a data point to the index}
\usage{
add_hnsw(ptr, data, label)
}
\arguments{
\item{data}{A vector of length (\code{dim}) to be added to the index.}

\item{label}{The corresponding scalar label for the data point.}

\item{index}{The index to which the data will be added.}
}
\value{
Returns \code{NULL}
}
\description{
This function appends a new data point to the existing kNN index.
}
\examples{
index <- create_index_hnsw(3, 16)
add_hnsw(index, c(1, 2, 3), 1)
}
