% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{add_knn}
\alias{add_knn}
\title{Add a data point to the index}
\usage{
add_knn(index, data)
}
\arguments{
\item{index}{The index to which the data will be added.}

\item{data}{The data point to be added to the index.}
}
\value{
An updated index with the new data point added.
}
\description{
This function appends a new data point to the existing kNN index.
}
\examples{
index <- create_index_knn(3)
index <- add_knn(index, c(1, 2, 3))
}
