% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Calculate Cosine Similarity}
\usage{
cosine_similarity(vec1, vec2)
}
\arguments{
\item{vec1}{The first vector.}

\item{vec2}{The second vector.}
}
\value{
The cosine similarity between the two input vectors.
}
\description{
This function computes the cosine similarity between two vectors.
}
