% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_index_hnsw}
\alias{create_index_hnsw}
\title{Create an HNSW index}
\usage{
create_index_hnsw(dim, max_size)
}
\arguments{
\item{dim}{The dimension of the feature space.}

\item{max_size}{The maximum capacity of the index}
}
\value{
A list containing an initialized index with an empty data matrix.
}
\description{
This function initializes an empty index for HNSW search with a specified dimension and max size.
}
