% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{create_index_knn}
\alias{create_index_knn}
\title{Create an index for k-Nearest Neighbors (kNN) search}
\usage{
create_index_knn(dim)
}
\arguments{
\item{dim}{The dimension of the feature space.}
}
\value{
A list containing an initialized index with an empty data matrix.
}
\description{
This function initializes an empty index for kNN search with a specified dimension.
}
\examples{
index <- create_index_knn(3)
}
