% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Calculate Euclidean Distance}
\usage{
euclidean_distance(vec1, vec2)
}
\arguments{
\item{vec1}{The first vector.}

\item{vec2}{The second vector.}
}
\value{
The Euclidean distance between the two input vectors.
}
\description{
This function computes the Euclidean distance between two vectors.
}
\examples{
euclidean_distance(c(1, 2, 3), c(4, 5, 6))
}
