% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_hnsw}
\alias{find_hnsw}
\title{Search for top K nearest neighbors}
\usage{
find_hnsw(ptr, q, k)
}
\arguments{
\item{ptr}{The index containing the data points.}

\item{q}{The query vector of length (\code{dim}) for which neighbors are to be found.}

\item{k}{The number of nearest neighbors to find.}
}
\value{
Indices of the k nearest neighbors.
}
\description{
This function finds the k nearest neighbors to a query vector within the index.
}
\examples{
index <- create_index_hnsw(3, 16)
add_hnsw(index, c(1, 2, 3), 1)
I <- find_hnsw(index, c(2, 3, 4), k = 1)
}
