% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{find_knn}
\alias{find_knn}
\title{Find k Nearest Neighbors (kNN)}
\usage{
find_knn(index, q, k = 5)
}
\arguments{
\item{index}{The index containing the data points.}

\item{q}{The query vector for which neighbors are to be found.}

\item{k}{The number of nearest neighbors to find (default is 5).}
}
\value{
Indices of the k nearest neighbors.
}
\description{
This function finds the k nearest neighbors to a query vector within the index.
}
\examples{
index <- create_index_knn(3)
index <- add_item_knn(index, c(1, 2, 3))
I <- find_knn(index, c(2, 3, 4), k = 2)
}
