% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnsw.R
\name{HNSWIndex}
\alias{HNSWIndex}
\title{R6 class representing a hierarchical navigable small worlds (HNSW) Index}
\description{
HNSW Index is used for nearest neighbor search based on provided data points.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{index}}{The dimension of the vectors to be indexed.}

\item{\code{dim}}{The dimension of the vectors to be indexed.}

\item{\code{size}}{The max capacity of the index.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HNSWIndex-new}{\code{HNSWIndex$new()}}
\item \href{#method-HNSWIndex-add}{\code{HNSWIndex$add()}}
\item \href{#method-HNSWIndex-find}{\code{HNSWIndex$find()}}
\item \href{#method-HNSWIndex-clone}{\code{HNSWIndex$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HNSWIndex-new"></a>}}
\if{latex}{\out{\hypertarget{method-HNSWIndex-new}{}}}
\subsection{Method \code{new()}}{
Create a new HNSW vector index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HNSWIndex$new(dim = NA, size = NA, M = 16, ef_construction = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dim}}{The dimension of the vectors to be indexed.}

\item{\code{size}}{The max capacity of the index.}

\item{\code{M}}{Parameter that defines the maximum number of outgoing connections in the graph.}

\item{\code{ef_construction}}{Parameter that controls speed/accuracy trade-off during the index construction.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HNSWIndex-add"></a>}}
\if{latex}{\out{\hypertarget{method-HNSWIndex-add}{}}}
\subsection{Method \code{add()}}{
Add a data point to the index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HNSWIndex$add(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A vector of length (\code{dim}) to be added to the index.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HNSWIndex-find"></a>}}
\if{latex}{\out{\hypertarget{method-HNSWIndex-find}{}}}
\subsection{Method \code{find()}}{
Search for top K nearest neighbors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HNSWIndex$find(q, k)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{The query vector of length (\code{dim})}

\item{\code{k}}{Number of results to return}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HNSWIndex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HNSWIndex-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HNSWIndex$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
