% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{KNNIndex}
\alias{KNNIndex}
\title{R6 class representing a k-Nearest Neighbors (kNN) Index}
\description{
kNN Index is used for nearest neighbor search based on provided data points.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KNNIndex-new}{\code{KNNIndex$new()}}
\item \href{#method-KNNIndex-add}{\code{KNNIndex$add()}}
\item \href{#method-KNNIndex-find}{\code{KNNIndex$find()}}
\item \href{#method-KNNIndex-clone}{\code{KNNIndex$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KNNIndex-new"></a>}}
\if{latex}{\out{\hypertarget{method-KNNIndex-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KNNIndex$new(dim = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KNNIndex-add"></a>}}
\if{latex}{\out{\hypertarget{method-KNNIndex-add}{}}}
\subsection{Method \code{add()}}{
Add a data point to the index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KNNIndex$add(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A vector of length (\code{dim}) to be added to the index.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KNNIndex-find"></a>}}
\if{latex}{\out{\hypertarget{method-KNNIndex-find}{}}}
\subsection{Method \code{find()}}{
Search for top K nearest neighbors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KNNIndex$find(q, k)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{The query of vector of length (\code{dim})}

\item{\code{k}}{Number of results to return}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KNNIndex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KNNIndex-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KNNIndex$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
