% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorstore.R
\name{VectorStore}
\alias{VectorStore}
\title{VectorStore class.}
\description{
VectorStore class for storing vector and non-vector data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{size}}{The current number of elements in the vector store.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VectorStore-new}{\code{VectorStore$new()}}
\item \href{#method-VectorStore-add}{\code{VectorStore$add()}}
\item \href{#method-VectorStore-find}{\code{VectorStore$find()}}
\item \href{#method-VectorStore-clone}{\code{VectorStore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-new"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new vector store.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$new(index = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{A vector index, either \code{HNSWIndex} or \code{KNNIndex}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-add"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-add}{}}}
\subsection{Method \code{add()}}{
Add a vector and a list of non-vector data to the store
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$add(vector, non_vector_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vector}}{A vector of length (\code{dim}) to be added to the index.}

\item{\code{non_vector_data}}{A list of non-vector data corresponding to the vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-find"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-find}{}}}
\subsection{Method \code{find()}}{
Search for top K nearest neighbors and return corresponding indexes and non-vector data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$find(q, k)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{The query vector of length (\code{dim}).}

\item{\code{k}}{Number of nearest neighbors to find.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A (q x p) data frame containing indexes of the k nearest neighbors and their corresponding non-vector data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
